Inherit from data types (#1161)
* Fixes to model spec

* Add the ability to resolved inherited data types

* Prevent circular references to data types

* Make resolved_config a virtual field only for gql

* Refactor index-model spec

* Add additional NOT query spec to index-model

* Ensure login/logout/auth don't use cache

* Add inherit_from dropdown

* WIP add the ability to show the resolved fields

* Use semantic-ui-css instead of semantic-ui-less

* Fixes to ActionSegment forms

* Fixes to ActionSegment and add build:watch to ui-core

* Fix eslint warning in ui-core

* Conditionally show the "show inherited fields button"

* Fixes to utils number/string and build watch

* Better validation for Type Config

* Add array support to type config

* Add test to ensure resolved version of data type is used on getViewForSpace

* Throw proper validation errors when version mismatch on inheritance

* Fix tests and don't show data types that are circular references in the dropdown

* Bump packages

* Fixes to build cleanup, ModelForm and Record Form

* Show resolved fields when editing a view

* Fix build cleanup script and getField type defs

+++ scripts/build-cleanup.sh
@@ -4,11 +4,11 @@ main() {
     local remove_only="$1"
     local cwd="$PWD";
     for name in $(yarn --silent lerna list --toposort); do
-        local package="${name#\@terascope}"
+        local package="${name#\@terascope\/}"
         if [ -z "$remove_only" ] || [ "$remove_only" == "dist" ]; then
             local dist_dir="$cwd/packages/$package/dist"
             if [ -d "$dist_dir" ]; then
-                rm -rf "$dist_dir"
+                rm -rf "${dist_dir:?}/*"
             fi
         fi
